+++ src/message.h
@@ -4,8 +4,8 @@
 #define KEY_LATITUDE 1
 #define KEY_TIMEZONE 2
   
-extern uint_fast16_t currentlong;
-extern uint_fast16_t currentlat;
+extern int currentlong;
+extern int currentlat;
 extern int_fast16_t timezone_offset;
 
 void message_init();
